requirements.txt
requests>=2.31.0
yt-dlp>=2024.8.6
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.6
demucs>=4.0.1
numpy<2
python-dotenv>=1.0.1
google-api-python-client>=2.150.0
google-auth-oauthlib>=1.2.1
google-cloud-storage>=2.18.2
redis>=5.0.0
